chore: Refactor webhook route to fetch user email from Clerk API
The code changes in `route.ts` add a new function `getUserEmail` that fetches the email address of a user from the Clerk API. This function is used in the `POST` function to retrieve the email address of a user when creating a new user. This change improves the reliability of the application by ensuring that the correct email address is associated with the user during the creation process.